fix(settings,theme): unify theme application across controls
Route the Appearance picker and cycle-theme hotkey through a shared theme
application hook. Keep settings cache, persisted values, and the DOM theme
class synchronized optimistically, with rollback when persistence fails.
