+++ src/lib/settings/queries.ts
@@ -1,9 +1,10 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { useMemo } from "react";
+import { useCallback, useMemo } from "react";
 import { PROVIDERS_REQUIRING_KEY } from "@/lib/ai/providers";
 import type { AiProviderId } from "@/lib/ai/types";
 import { getSecret } from "@/lib/git/api";
 import { repoIdentity } from "@/lib/git/repo-identity";
+import { commitTheme, type ThemeSetting } from "@/lib/theme";
 import {
   type AppSettings,
   addRecentRepo,
@@ -102,6 +103,34 @@ export function useSaveSettings() {
     onSuccess: () =>
       queryClient.invalidateQueries({ queryKey: settingsKeys.settings }),
   });
+}
+
+/**
+ * Apply a theme change from any entry point — the Appearance picker and the
+ * `cycle-theme` command both go through this, so the two paths can't drift. It
+ * optimistically patches the settings cache (so a bound `<Select>` reflects the
+ * choice immediately, not a refetch later), persists it, applies the DOM class,
+ * and rolls all three back if the store write throws. `useSaveSettings`'s
+ * success-invalidate reconciles the cache on the happy path.
+ */
+export function useApplyTheme() {
+  const queryClient = useQueryClient();
+  const saveSettings = useSaveSettings();
+  return useCallback(
+    (current: AppSettings, next: ThemeSetting) => {
+      if (next === current.theme) return;
+      const updated = { ...current, theme: next };
+      queryClient.setQueryData(settingsKeys.settings, updated);
+      commitTheme(next);
+      saveSettings.mutate(updated, {
+        onError: () => {
+          queryClient.setQueryData(settingsKeys.settings, current);
+          commitTheme(current.theme);
+        },
+      });
+    },
+    [queryClient, saveSettings],
+  );
 }
 
 export function useAddRecentRepo() {
